+++ HabboHotel/Users/IHabboRuntimeInitializer.cs
@@ -1,0 +1,7 @@
+namespace Plus.HabboHotel.Users;
+
+public interface IHabboRuntimeInitializer
+{
+    void EnsureVisualComponents(Habbo habbo);
+    void EnsureProcessComponent(Habbo habbo);
+}